Consolidate ObservationMapper version and ObsBuilder versions
