+++ api/src/main/resources/moduleApplicationContext.xml
@@ -320,6 +320,12 @@
         <constructor-arg name="observationTypeMatcher" ref="observationTypeMatcher"/>
     </bean>
 
+    <bean id="observationMapper" class="org.openmrs.module.emrapi.encounter.ObservationMapper">
+        <constructor-arg name="conceptMapper" ref="conceptMapper"/>
+        <constructor-arg name="drugMapper" ref="drugMapper"/>
+        <constructor-arg name="userMapper" ref="userMapper"/>
+    </bean>
+
     <bean id="conceptMapper" class="org.openmrs.module.emrapi.encounter.ConceptMapper"/>
 
     <bean id="diagnosisMapper" class="org.openmrs.module.emrapi.encounter.DiagnosisMapper">
